requirements-cloud.txt
# Минимальные зависимости для облачного деплоя (без Rust компиляции)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1